dom/messagechannel/tests/test_wasm.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>MessagePort + wasm</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<script id='myWorker' type='text/worker'>
self.onmessage = e => {
  e.ports[0].onmessage = ee => postMessage(ee.data);
};
</script>

<script type="application/javascript">

let wasmModule;

function create_wasmModule() {
  info("Checking if we can play with WebAssembly...");

  if (!SpecialPowers.Cu.getJSTestingFunctions().wasmIsSupported()) {
    next();
    return;
  }

  ok(WebAssembly, "WebAssembly object should exist");
  ok(WebAssembly.compile, "WebAssembly.compile function should exist");

  const wasmTextToBinary = SpecialPowers.unwrap(SpecialPowers.Cu.getJSTestingFunctions().wasmTextToBinary);
  const fooModuleCode = wasmTextToBinary(`(module
    (func $foo (result i32) (i32.const 42))
    (export "foo" $foo)
  )`);

  WebAssembly.compile(fooModuleCode).then(m => {
    ok(m instanceof WebAssembly.Module, "The WasmModule has been compiled.");
    wasmModule = m;
    next();
  }, () => {
    ok(false, "The compilation of the wasmModule failed.");
  });
}

function test_simple() {
  info("MessagePort without disentangling");

  if (!wasmModule) {
    next();
    return;
  }

  const mc = new MessageChannel();
  mc.port1.postMessage({ wasmModule });
  mc.port2.onmessage = e => {
    ok(e.data.wasmModule instanceof WebAssembly.Module, "WasmModule received");
    next();
  };
}

function test_worker() {
  info("MessagePort sent to a worker");

  if (!wasmModule) {
    next();
    return;
  }

  const mc = new MessageChannel();
  mc.port1.postMessage({ wasmModule });

  const blob = new Blob([document.getElementById('myWorker').textContent]);
  const w = new Worker(URL.createObjectURL(blob));
  w.postMessage(42, [ mc.port2 ]);
  w.onmessage = e => {
    ok(e.data.wasmModule instanceof WebAssembly.Module, "WasmModule received");
    next();
  };
}

const tests = [
  create_wasmModule,

  test_simple,
  test_worker,
];

function next() {
  if (tests.length === 0) {
    SimpleTest.finish();
    return;
  }

  const test = tests.shift();
  test();
}

SimpleTest.waitForExplicitFinish();
next();

</script>
</body>
</html>
